+++ packages/cursorless-engine/src/core/Snippets.ts
@@ -1,4 +1,4 @@
-import type { Snippet, SnippetMap, TextEditor } from "@cursorless/common";
+import type { TextEditor } from "@cursorless/common";
 
 /**
  * Handles all cursorless snippets, including core, third-party and
@@ -6,28 +6,6 @@ import type { Snippet, SnippetMap, TextEditor } from "@cursorless/common";
  * name.
  */
 export interface Snippets {
-  updateUserSnippets(): Promise<void>;
-
-  /**
-   * Allows extensions to register third-party snippets.  Calling this function
-   * twice with the same extensionId will replace the older snippets.
-   *
-   * Note that third-party snippets take precedence over core snippets, but
-   * user snippets take precedence over both.
-   * @param extensionId The id of the extension registering the snippets.
-   * @param snippets The snippets to be registered.
-   */
-  registerThirdPartySnippets(extensionId: string, snippets: SnippetMap): void;
-
-  /**
-   * Looks in merged collection of snippets for a snippet with key
-   * `snippetName`. Throws an exception if the snippet of the given name could
-   * not be found
-   * @param snippetName The name of the snippet to look up
-   * @returns The named snippet
-   */
-  getSnippetStrict(snippetName: string): Snippet;
-
   /**
    * Opens a new snippet file
    * @param snippetName The name of the snippet
@@ -36,6 +14,6 @@ export interface Snippets {
    */
   openNewSnippetFile(
     snippetName: string,
-    directory?: string,
+    directory: string,
   ): Promise<TextEditor>;
 }
